ch(route): properly define Agent route
This commit implements the full route for agents/employees. The fun-
ctioinality is set up to accept image uploads and directly uploading
same to cloudinary without the need to momentarily store same in the
local file storage. This is to handle instances where the host may not
allow such local stoorage.